requirements.txt
pandas
numpy
scipy
requests
beautifulsoup4
python-dotenv
pyyaml